requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
mysql-connector-python==9.4.0
nba_api==1.10.0
numpy==2.3.1
pandas==2.3.1
playwright==1.54.0
pyee==13.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
